+++ src/common/regions/__tests__/inline-hash.spec.ts
@@ -1,75 +1,77 @@
 import { REGION_GROUP_MATCHER } from '../parser';
-import { HTML_MATCHER } from './html';
+import { INLINE_HASH_MATCHER } from '../inline-hash';
 
 const regionName = 'RegionName';
 const regionName2 = 'RegionName2';
 
-const regionContents = `	<div>
-		<p>Some text with a <a href='#'>Link</a>!</p>
-	</div>`;
+const regionContents = `	some:
+		setting:
+			# This is for setting a message
+			property: 'Hey there!'
+			otherProperty: true`;
 
-const region = `	<!-- @start-region RegionName -->
+const region = `	# @start-region RegionName
 ${regionContents}
-	<!-- @end-region RegionName -->`;
+	# @end-region RegionName`;
 
-describe('html region-matcher', () => {
+describe('inline-hash region-matcher', () => {
 	it('should match start annotations', () => {
 		let matches: RegExpExecArray | null;
 
-		const regionMatcher = HTML_MATCHER.factory(regionName);
+		const regionMatcher = INLINE_HASH_MATCHER.factory(regionName);
 		const startMatcher = new RegExp(regionMatcher.regionStartMatcher);
 
-		matches = startMatcher.exec(`<!-- @start-region ${regionName} -->`);
+		matches = startMatcher.exec(`# @start-region ${regionName}`);
 		expect(matches).not.toBeNull();
 
-		matches = startMatcher.exec(`<!--@start-region ${regionName}-->`);
+		matches = startMatcher.exec(`#@start-region ${regionName}`);
 		expect(matches).not.toBeNull();
 
-		matches = startMatcher.exec(`<!-- @start-region ${regionName2} -->`);
+		matches = startMatcher.exec(`# @start-region ${regionName2}`);
 		expect(matches).toBeNull();
 	});
 
 	it('should match end annotations', () => {
 		let matches: RegExpExecArray | null;
 
-		const regionMatcher = HTML_MATCHER.factory(regionName);
+		const regionMatcher = INLINE_HASH_MATCHER.factory(regionName);
 		const endMatcher = new RegExp(regionMatcher.regionEndMatcher);
 
-		matches = endMatcher.exec(`<!-- @end-region ${regionName} -->`);
+		matches = endMatcher.exec(`# @end-region ${regionName} `);
 		expect(matches).not.toBeNull();
 
-		matches = endMatcher.exec(`<!--@end-region ${regionName}-->`);
+		matches = endMatcher.exec(`#@end-region ${regionName}`);
 		expect(matches).not.toBeNull();
 
-		matches = endMatcher.exec(`<!-- @end-region ${regionName2} -->`);
+		matches = endMatcher.exec(`# @end-region ${regionName2}`);
 		expect(matches).toBeNull();
 	});
 
 	it('should match all region comments', () => {
 		let matches: RegExpMatchArray | null;
 
-		const regionMatcher = HTML_MATCHER.regionCommentMatcher;
+		const regionMatcher = INLINE_HASH_MATCHER.regionCommentMatcher;
 
-		matches = `<!-- @end-region ${regionName} -->`.match(regionMatcher);
+		matches = `# @end-region ${regionName} `.match(regionMatcher);
 		expect(matches).not.toBeNull();
 
-		matches = `<!--@end-region ${regionName2}-->`.match(regionMatcher);
+		matches = `#@end-region ${regionName2}`.match(regionMatcher);
 		expect(matches).not.toBeNull();
 
-		matches = `<!-- @start-region ${regionName} -->`.match(regionMatcher);
+		matches = `# @start-region ${regionName} `.match(regionMatcher);
 		expect(matches).not.toBeNull();
 
-		matches = `<!--@start-region ${regionName2}-->`.match(regionMatcher);
+		matches = `#@start-region ${regionName2}`.match(regionMatcher);
 		expect(matches).not.toBeNull();
 
-		matches = `<!-- @middle-region ${regionName} -->`.match(regionMatcher);
+		matches = `# @middle-region ${regionName} `.match(regionMatcher);
 		expect(matches).toBeNull();
 	});
 
 	it('should match full region and retreive contents', () => {
 		let matches: RegExpExecArray | null;
 
-		const regionMatcher = HTML_MATCHER.factory(regionName);
+		const regionMatcher = INLINE_HASH_MATCHER.factory(regionName);
 		const regionRegExp = new RegExp(
 			`${regionMatcher.regionStartMatcher}${REGION_GROUP_MATCHER}${regionMatcher.regionEndMatcher}`,
 			'gm'
