+++ src/examples/list.block.ts
@@ -2,7 +2,7 @@ import fetch from 'node-fetch';
 import { VNode } from '@dojo/framework/core/interfaces';
 import { select } from '@dojo/framework/testing/support/selector';
 
-import { fromMarkdown, toDNodes } from './compile';
+import markdown from '../common/markdown';
 
 const README_URL = 'https://raw.githubusercontent.com/dojo/examples/master/README.md';
 
@@ -32,18 +32,18 @@ export default async function(): Promise<ExampleMeta[]> {
 			.map((value) => value.trim())
 			.filter((value) => value);
 		const exampleName = data[1].replace(/((^\[Link\]\(\.\/)|(\)$))/g, '');
-		const demoLink = select('a', toDNodes(fromMarkdown(data[2], {})));
+		const demoLink = select('a', markdown(data[2]));
 		let demoUrl = '';
 		if (demoLink && demoLink.length === 1) {
 			demoUrl = (demoLink[0].properties as any).href;
 		}
 
 		return {
-			[keys[0]]: toDNodes(fromMarkdown(data[0], {})),
-			[keys[1]]: toDNodes(fromMarkdown(data[1], {})),
+			[keys[0]]: markdown(data[0]),
+			[keys[1]]: markdown(data[1]),
 			[keys[2]]: demoUrl,
 			[keys[3]]: data.length === keys.length,
-			[keys[4]]: toDNodes(fromMarkdown(data.length === keys.length ? data[4] : data[3], {})),
+			[keys[4]]: markdown(data.length === keys.length ? data[4] : data[3]),
 			exampleName
 		};
 	});
